+++ samples/cmp-worker-sample-compose-store/src/commonMain/kotlin/io/github/mobilebytelabs/worker/sample/composestore/domain/Article.kt
@@ -1,0 +1,8 @@
+package io.github.mobilebytelabs.worker.sample.composestore.domain
+
+data class Article(
+    val id: String,
+    val title: String,
+    val body: String,
+    val fetchedAtEpochMs: Long,
+)